app/requirements.txt
asgiref
channels
channels-redis
coreapi
coreschema
cryptography
daphne
dj-database-url
Django
django-cloudinary-storage
django-cors-headers
django-phonenumber-field
django-redis
django-filter
django-rest-framework
django-templated-mail
djangorestframework
djangorestframework-simplejwt
djoser
drf-yasg
gunicorn
h2
hyperlink
Jinja2
MarkupSafe
phonenumbers
psycopg2-binary
PyJWT
Pillow
pyOpenSSL
pyparsing
python-decouple
python3-openid
pytz
redis
requests
requests-oauthlib
rest-framework-simplejwt
social-auth-app-django
social-auth-core
Twisted[tls,http2]
whitenoise
uvicorn
